terminacion de proyecto

+++ src/app/component/clothes-list/clothes-list.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit, HostBinding } from '@angular/core';
 import { Router, NavigationEnd } from '@angular/router';
 
 import { ClothesService } from '../../services/clothes.service'
+import { Observable, map } from 'rxjs';
 
 @Component({
   selector: 'app-clothes-list',
@@ -13,8 +14,11 @@ export class ClothesListComponent implements OnInit {
   @HostBinding('class') classes = 'row';
 
   clothes: any = [];
-  showButtons = false;
+  cantExistente: any = [];
 
+  showButtons = false;
+  itemId: string = "1"; 
+  cantidad: number = 0;
 
   constructor(private clothesService: ClothesService, private router: Router) { }
 
@@ -32,13 +36,54 @@ export class ClothesListComponent implements OnInit {
     );
   }
 
-  deleteClothes(id: string){
-    console.log(id);
-    this.clothesService.deleteClothe(id).subscribe(
-      res => {
-        console.log(res);
-        this.getClothes();
-      }, err => console.log(err)
-    )
+  getClothesExisting(id: string): Observable<number> {
+    return this.clothesService.getClothesExisting(id).pipe(
+      map(res => {
+        this.cantExistente = res;
+        return this.cantExistente[0].cantidad_existente;
+      })
+    );
+  }
+
+  removeQuantityExisting(id: string) {
+    //consulta cantidades existentes
+    this.getClothesExisting(id).subscribe(
+      cantExistentes => {
+        this.cantidad = cantExistentes as number;
+        // Aquí puedes usar el valor de cantExistentes como necesites
+        if (this.cantidad > 0) {
+          this.clothesService.removeQuantityExisting(id).subscribe(
+              res => {
+                console.log(res);
+                this.getClothes();
+                this.getClothesExisting(id).subscribe(
+                  nuevaCantidad => {
+                    if (nuevaCantidad === 0) {
+                      // Si la cantidad es 0 después de la eliminación, elimina automáticamente el objeto
+                      this.clothesService.deleteClothe(id).subscribe(
+                        deleteRes => {
+                          console.log(deleteRes);
+                          this.getClothes();
+                        },
+                        deleteErr => console.error(deleteErr)
+                      );
+                    }
+                  }, getCantidadError => {
+                    console.error('Error al obtener la cantidad existente después de la eliminación:', getCantidadError);
+                  }
+                );
+              },
+             err => console.log(err)
+            );
+        } 
+      },
+      error => {
+        console.error('Error al obtener la cantidad existente:', error);
+      }
+    );
+    
+    
+    
+    
   }
 }
